@@ -17,7 +17,6 @@ license = MIT
 classifiers =
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
@@ -26,7 +25,7 @@ long_description_content_type = text/x-rst
 
 [options]
 # When updating this value, make sure our CI matrix includes a matching minimum version.
-python_requires = >=3.6.7
+python_requires = >=3.7.1
 install_requires =
     zope.interface >= 4.4.2
     constantly >= 15.1
@@ -72,7 +71,7 @@ dev =
     coverage >= 6b1, <7
 
 tls =
-    pyopenssl >= 16.0.0
+    pyopenssl >= 21.0.0
     # service_identity 18.1.0 added support for validating IP addresses in
     # certificate subjectAltNames
     service_identity >= 18.1.0
@@ -120,6 +119,10 @@ windows_platform =
 
 osx_platform =
     %(macos_platform)s
+
+gtk_platform =
+    pygobject
+    %(all_non_platform)s
 
 mypy =
     mypy==0.930
