@@ -11,25 +11,12 @@ import binascii
 import errno
 import sys
 from base64 import decodebytes
-from typing import BinaryIO, Callable, Iterator
-
-try:
-    import pwd as _pwd
-except ImportError:
-    pwd = None
-else:
-    pwd = _pwd
-
-try:
-    import spwd as _spwd
-except ImportError:
-    spwd = None
-else:
-    spwd = _spwd
+from typing import IO, Callable, Iterable, Iterator, Mapping, Optional, Tuple, cast
 
 from zope.interface import Interface, implementer, providedBy
 
 from incremental import Version
+from typing_extensions import Literal, Protocol
 
 from twisted.conch import error
 from twisted.conch.ssh import keys
@@ -47,21 +34,97 @@ from twisted.python.util import runAsEffectiveUser
 _log = Logger()
 
 
-def _pwdGetByName(username):
+class UserRecord(Tuple[str, str, int, int, str, str, str]):
+    """
+    A record in a UNIX-style password database. See L{pwd} for field details.
+
+    This corresponds to the undocumented type L{pwd.struct_passwd}, but lacks named
+    field accessors.
+    """
+
+    @property
+    def pw_dir(self) -> str:
+        ...
+
+
+class UserDB(Protocol):
+    """
+    A database of users by name, like the stdlib L{pwd} module.
+
+    See L{twisted.python.fakepwd} for an in-memory implementation.
+    """
+
+    def getpwnam(self, username: str) -> UserRecord:
+        """
+        Lookup a user record by name.
+
+        @raises KeyError: when no such user exists
+        """
+
+
+pwd: Optional[UserDB]
+try:
+    import pwd as _pwd
+except ImportError:
+    pwd = None
+else:
+    pwd = cast(UserDB, _pwd)
+
+
+try:
+    import spwd as _spwd
+except ImportError:
+    spwd = None
+else:
+    spwd = _spwd
+
+
+class CryptedPasswordRecord(Protocol):
+    """
+    A sequence where the item at index 1 may be a crypted password.
+
+    Both L{pwd.struct_passwd} and L{spwd.struct_spwd} conform to this protocol.
+    """
+
+    def __getitem__(self, index: Literal[1]) -> str:
+        """
+        Get the crypted password.
+        """
+
+
+def _lookupUser(userdb: UserDB, username: bytes) -> UserRecord:
+    """
+    Lookup a user by name in a L{pwd}-style database.
+
+    @param userdb: The user database.
+
+    @param username: Identifying name in bytes. This will be decoded according
+    to the filesystem encoding, as the L{pwd} module does internally.
+
+    @raises KeyError: when the user doesn't exist
+    """
+    return userdb.getpwnam(username.decode(sys.getfilesystemencoding()))
+
+
+def _pwdGetByName(username: str) -> Optional[CryptedPasswordRecord]:
     """
     Look up a user in the /etc/passwd database using the pwd module.  If the
     pwd module is not available, return None.
 
     @param username: the username of the user to return the passwd database
         information for.
-    @type username: L{str}
+
+    @returns: A L{pwd.struct_passwd}, where field 1 may contain a crypted
+        password, or L{None} when the L{pwd} database is unavailable.
+
+    @raises KeyError: when no such user exists
     """
     if pwd is None:
         return None
-    return pwd.getpwnam(username)
+    return cast(CryptedPasswordRecord, pwd.getpwnam(username))
 
 
-def _shadowGetByName(username):
+def _shadowGetByName(username: str) -> Optional[CryptedPasswordRecord]:
     """
     Look up a user in the /etc/shadow database using the spwd module. If it is
     not available, return L{None}.
@@ -69,12 +132,17 @@ def _shadowGetByName(username):
     @param username: the username of the user to return the shadow database
         information for.
     @type username: L{str}
+
+    @returns: A L{spwd.struct_spwd}, where field 1 may contain a crypted
+        password, or L{None} when the L{spwd} database is unavailable.
+
+    @raises KeyError: when no such user exists
     """
     if spwd is not None:
         f = spwd.getspnam
     else:
         return None
-    return runAsEffectiveUser(0, 0, f, username)
+    return cast(CryptedPasswordRecord, runAsEffectiveUser(0, 0, f, username))
 
 
 @implementer(ICredentialsChecker)
@@ -84,7 +152,7 @@ class UNIXPasswordDatabase:
     databases of a compatible format.
 
     @ivar _getByNameFunctions: a C{list} of functions which are called in order
-        to valid a user.  The default value is such that the C{/etc/passwd}
+        to validate a user.  The default value is such that the C{/etc/passwd}
         database will be tried first, followed by the C{/etc/shadow} database.
     """
 
@@ -127,7 +195,7 @@ class SSHPublicKeyDatabase:
 
     credentialInterfaces = (ISSHPrivateKey,)
 
-    _userdb = pwd
+    _userdb: UserDB = cast(UserDB, pwd)
 
     def requestAvatarId(self, credentials):
         d = defer.maybeDeferred(self.checkKey, credentials)
@@ -186,7 +254,7 @@ class SSHPublicKeyDatabase:
 
         @return: A list of L{FilePath} instances to files with the authorized keys.
         """
-        pwent = self._userdb.getpwnam(credentials.username)
+        pwent = _lookupUser(self._userdb, credentials.username)
         root = FilePath(pwent.pw_dir).child(".ssh")
         files = ["authorized_keys", "authorized_keys2"]
         return [root.child(f) for f in files]
@@ -196,7 +264,7 @@ class SSHPublicKeyDatabase:
         Retrieve files containing authorized keys and check against user
         credentials.
         """
-        ouid, ogid = self._userdb.getpwnam(credentials.username)[2:4]
+        ouid, ogid = _lookupUser(self._userdb, credentials.username)[2:4]
         for filepath in self.getAuthorizedKeysFiles(credentials):
             if not filepath.exists():
                 continue
@@ -339,7 +407,7 @@ class IAuthorizedKeysDB(Interface):
 
 
 def readAuthorizedKeyFile(
-    fileobj: BinaryIO, parseKey: Callable[[bytes], keys.Key] = keys.Key.fromString
+    fileobj: IO[bytes], parseKey: Callable[[bytes], keys.Key] = keys.Key.fromString
 ) -> Iterator[keys.Key]:
     """
     Reads keys from an authorized keys file.  Any non-comment line that cannot
@@ -367,7 +435,9 @@ def readAuthorizedKeyFile(
                 )
 
 
-def _keysFromFilepaths(filepaths, parseKey):
+def _keysFromFilepaths(
+    filepaths: Iterable[FilePath], parseKey: Callable[[bytes], keys.Key]
+) -> Iterable[keys.Key]:
     """
     Helper function that turns an iterable of filepaths into a generator of
     keys.  If any file cannot be read, a message is logged but it is
@@ -381,7 +451,6 @@ def _keysFromFilepaths(filepaths, parseKey):
     @type parseKey: L{callable}
 
     @return: generator of L{twisted.conch.ssh.keys.Key}
-    @rtype: generator
 
     @since: 15.0
     """
@@ -403,18 +472,22 @@ class InMemorySSHKeyDB:
     @since: 15.0
     """
 
-    def __init__(self, mapping):
+    def __init__(self, mapping: Mapping[bytes, Iterable[keys.Key]]) -> None:
         """
         Initializes a new L{InMemorySSHKeyDB}.
 
         @param mapping: mapping of usernames to iterables of
             L{twisted.conch.ssh.keys.Key}s
-        @type mapping: L{dict}
 
         """
         self._mapping = mapping
 
-    def getAuthorizedKeys(self, username):
+    def getAuthorizedKeys(self, username: bytes) -> Iterable[keys.Key]:
+        """
+        Look up the authorized keys for a user.
+
+        @param username: Name of the user
+        """
         return self._mapping.get(username, [])
 
 
@@ -429,27 +502,34 @@ class UNIXAuthorizedKeysFiles:
     @since: 15.0
     """
 
-    def __init__(self, userdb=None, parseKey=keys.Key.fromString):
+    _userdb: UserDB
+
+    def __init__(
+        self,
+        userdb: Optional[UserDB] = None,
+        parseKey: Callable[[bytes], keys.Key] = keys.Key.fromString,
+    ):
         """
         Initializes a new L{UNIXAuthorizedKeysFiles}.
 
         @param userdb: access to the Unix user account and password database
-            (default is the Python module L{pwd})
-        @type userdb: L{pwd}-like object
+            (default is the Python module L{pwd}, if available)
 
         @param parseKey: a callable that takes a string and returns a
             L{twisted.conch.ssh.keys.Key}, mainly to be used for testing.  The
             default is L{twisted.conch.ssh.keys.Key.fromString}.
-        @type parseKey: L{callable}
         """
-        self._userdb = userdb
-        self._parseKey = parseKey
-        if userdb is None:
+        if userdb is not None:
+            self._userdb = userdb
+        elif pwd is not None:
             self._userdb = pwd
+        else:
+            raise ValueError("No pwd module found, and no userdb argument passed.")
+        self._parseKey = parseKey
 
-    def getAuthorizedKeys(self, username):
+    def getAuthorizedKeys(self, username: bytes) -> Iterable[keys.Key]:
         try:
-            passwd = self._userdb.getpwnam(username)
+            passwd = _lookupUser(self._userdb, username)
         except KeyError:
             return ()
 
@@ -472,12 +552,11 @@ class SSHPublicKeyChecker:
 
     credentialInterfaces = (ISSHPrivateKey,)
 
-    def __init__(self, keydb):
+    def __init__(self, keydb: IAuthorizedKeysDB) -> None:
         """
         Initializes a L{SSHPublicKeyChecker}.
 
         @param keydb: a provider of L{IAuthorizedKeysDB}
-        @type keydb: L{IAuthorizedKeysDB} provider
         """
         self._keydb = keydb
 
