@@ -135,7 +135,7 @@ def generateRSAkey(options):
     from cryptography.hazmat.primitives.asymmetric import rsa
 
     if not options["bits"]:
-        options["bits"] = 1024
+        options["bits"] = 2048
     keyPrimitive = rsa.generate_private_key(
         key_size=int(options["bits"]),
         public_exponent=65537,
