@@ -1273,9 +1273,9 @@ class OpenSSLCertificateOptions:
     @type _cipherString: L{unicode}
 
     @ivar _defaultMinimumTLSVersion: The default TLS version that will be
-        negotiated. This should be a "safe default", with wide client and
+        negotiated.  This should be a "safe default", with wide client and
         server support, vs an optimally secure one that excludes a large number
-        of users. As of late 2016, TLSv1.0 is that safe default.
+        of users.  As of May 2022, TLSv1.2 is that safe default.
     @type _defaultMinimumTLSVersion: L{TLSVersion} constant
     """
 
@@ -1285,7 +1285,7 @@ class OpenSSLCertificateOptions:
 
     _OP_NO_TLSv1_3 = _tlsDisableFlags[TLSVersion.TLSv1_3]
 
-    _defaultMinimumTLSVersion = TLSVersion.TLSv1_0
+    _defaultMinimumTLSVersion = TLSVersion.TLSv1_2
 
     @_mutuallyExclusiveArguments(
         [
@@ -1331,11 +1331,11 @@ class OpenSSLCertificateOptions:
         @param method: Deprecated, use a combination of
             C{insecurelyLowerMinimumTo}, C{raiseMinimumTo}, or
             C{lowerMaximumSecurityTo} instead.  The SSL protocol to use, one of
-            C{SSLv23_METHOD}, C{SSLv2_METHOD}, C{SSLv3_METHOD}, C{TLSv1_METHOD}
-            (or any other method constants provided by pyOpenSSL).  By default,
-            a setting will be used which allows TLSv1.0, TLSv1.1, and TLSv1.2.
-            Can not be used with C{insecurelyLowerMinimumTo},
-            C{raiseMinimumTo}, or C{lowerMaximumSecurityTo}
+            C{TLS_METHOD}, C{TLSv1_2_METHOD}, or C{TLSv1_2_METHOD} (or any
+            future method constants provided by pyOpenSSL).  By default, a
+            setting will be used which allows TLSv1.2 and TLSv1.3.  Can not be
+            used with C{insecurelyLowerMinimumTo}, C{raiseMinimumTo}, or
+            C{lowerMaximumSecurityTo}.
 
         @param verify: Please use a C{trustRoot} keyword argument instead,
             since it provides the same functionality in a less error-prone way.
@@ -1489,7 +1489,7 @@ class OpenSSLCertificateOptions:
         self._mode = SSL.MODE_RELEASE_BUFFERS
 
         if method is None:
-            self.method = SSL.SSLv23_METHOD
+            self.method = SSL.TLS_METHOD
 
             if raiseMinimumTo:
                 if lowerMaximumSecurityTo and raiseMinimumTo > lowerMaximumSecurityTo:
@@ -1722,13 +1722,13 @@ class OpenSSLCipher:
 @lru_cache(maxsize=32)
 def _expandCipherString(cipherString, method, options):
     """
-    Expand C{cipherString} according to C{method} and C{options} to a tuple
-    of explicit ciphers that are supported by the current platform.
+    Expand C{cipherString} according to C{method} and C{options} to a tuple of
+    explicit ciphers that are supported by the current platform.
 
     @param cipherString: An OpenSSL cipher string to expand.
     @type cipherString: L{unicode}
 
-    @param method: An OpenSSL method like C{SSL.TLSv1_METHOD} used for
+    @param method: An OpenSSL method like C{SSL.TLS_METHOD} used for
         determining the effective ciphers.
 
     @param options: OpenSSL options like C{SSL.OP_NO_SSLv3} ORed together.
@@ -1810,7 +1810,7 @@ class OpenSSLAcceptableCiphers:
         return cls(
             _expandCipherString(
                 nativeString(cipherString),
-                SSL.SSLv23_METHOD,
+                SSL.TLS_METHOD,
                 SSL.OP_NO_SSLv2 | SSL.OP_NO_SSLv3,
             )
         )
