@@ -14,8 +14,8 @@ from abc import ABC, abstractmethod
 from asyncio import AbstractEventLoop, Future, iscoroutine
 from enum import Enum
 from functools import wraps
-from sys import exc_info, version_info
-from types import GeneratorType, MappingProxyType
+from sys import exc_info
+from types import CoroutineType, GeneratorType, MappingProxyType
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -39,7 +39,7 @@ from typing import (
 
 import attr
 from incremental import Version
-from typing_extensions import Literal
+from typing_extensions import Literal, Protocol
 
 from twisted.internet.interfaces import IDelayedCall, IReactorTime
 from twisted.logger import Logger
@@ -47,6 +47,12 @@ from twisted.python import lockfile
 from twisted.python.compat import _PYPY, cmp, comparable
 from twisted.python.deprecate import deprecated, warnAboutFunction
 from twisted.python.failure import Failure, _extraneous
+
+
+class _Context(Protocol):
+    def run(self, f: Callable[..., object], *args: object, **kwargs: object) -> object:
+        ...
+
 
 try:
     from contextvars import copy_context as __copy_context
@@ -171,13 +177,22 @@ def maybeDeferred(
     f: Callable[..., _T], *args: object, **kwargs: object
 ) -> "Deferred[_T]":
     """
-    Invoke a function that may or may not return a L{Deferred}.
+    Invoke a function that may or may not return a L{Deferred} or coroutine.
 
-    Call the given function with the given arguments.  If the returned
-    object is a L{Deferred}, return it.  If the returned object is a L{Failure},
-    wrap it with L{fail} and return it.  Otherwise, wrap it in L{succeed} and
-    return it.  If an exception is raised, convert it to a L{Failure}, wrap it
-    in L{fail}, and then return it.
+    Call the given function with the given arguments.  Then:
+
+      - If the returned object is a L{Deferred}, return it.
+
+      - If the returned object is a L{Failure}, wrap it with L{fail} and
+        return it.
+
+      - If the returned object is a L{types.CoroutineType}, wrap it with
+        L{Deferred.fromCoroutine} and return it.
+
+      - Otherwise, wrap it in L{succeed} and return it.
+
+      - If an exception is raised, convert it to a L{Failure}, wrap it in
+        L{fail}, and then return it.
 
     @param f: The callable to invoke
     @param args: The arguments to pass to C{f}
@@ -195,6 +210,25 @@ def maybeDeferred(
         return result
     elif isinstance(result, Failure):
         return fail(result)
+    elif type(result) is CoroutineType:
+        # A note on how we identify this case ...
+        #
+        # inspect.iscoroutinefunction(f) should be the simplest and easiest
+        # way to determine if we want to apply coroutine handling.  However,
+        # the value may be returned by a regular function that calls a
+        # coroutine function and returns its result.  It would be confusing if
+        # cases like this led to different handling of the coroutine (even
+        # though it is a mistake to have a regular function call a coroutine
+        # function to return its result - doing so immediately destroys a
+        # large part of the value of coroutine functions: that they can only
+        # have a coroutine result).
+        #
+        # There are many ways we could inspect ``result`` to determine if it
+        # is a "coroutine" but most of these are mistakes.  The goal is only
+        # to determine whether the value came from ``async def`` or not
+        # because these are the only values we're trying to handle with this
+        # case.  Such values always have exactly one type: CoroutineType.
+        return Deferred.fromCoroutine(result)
     else:
         return succeed(result)
 
@@ -1616,6 +1650,7 @@ def _inlineCallbacks(
         Coroutine[Deferred[_T], object, None],
     ],
     status: _CancellationStatus,
+    context: _Context,
 ) -> None:
     """
     Carry out the work of L{inlineCallbacks}.
@@ -1634,6 +1669,8 @@ def _inlineCallbacks(
         decorated with C{@}L{inlineCallbacks}
 
     @param status: a L{_CancellationStatus} tracking the current status of C{gen}
+
+    @param context: the contextvars context to run `gen` in
     """
     # This function is complicated by the need to prevent unbounded recursion
     # arising from repeatedly yielding immediately ready deferreds.  This while
@@ -1643,20 +1680,17 @@ def _inlineCallbacks(
     # waiting for result?  # result
     waiting: List[Any] = [True, None]
 
-    # Get the current contextvars Context object.
-    current_context = _copy_context()
-
     while 1:
         try:
             # Send the last result back as the result of the yield expression.
             isFailure = isinstance(result, Failure)
 
             if isFailure:
-                result = current_context.run(
+                result = context.run(
                     cast(Failure, result).throwExceptionIntoGenerator, gen
                 )
             else:
-                result = current_context.run(gen.send, result)
+                result = context.run(gen.send, result)
         except StopIteration as e:
             # fell off the end, or "return" statement
             status.deferred.callback(getattr(e, "value", None))
@@ -1680,11 +1714,7 @@ def _inlineCallbacks(
             appCodeTrace = traceback.tb_next
             assert appCodeTrace is not None
 
-            if version_info < (3, 7):
-                # The contextvars backport and our no-op shim add an extra frame.
-                appCodeTrace = appCodeTrace.tb_next
-                assert appCodeTrace is not None
-            elif _PYPY:
+            if _PYPY:
                 # PyPy as of 3.7 adds an extra frame.
                 appCodeTrace = appCodeTrace.tb_next
                 assert appCodeTrace is not None
@@ -1746,7 +1776,7 @@ def _inlineCallbacks(
                     waiting[0] = False
                     waiting[1] = r
                 else:
-                    current_context.run(_inlineCallbacks, r, gen, status)
+                    _inlineCallbacks(r, gen, status, context)
 
             result.addBoth(gotResult)
             if waiting[0]:
@@ -1811,7 +1841,7 @@ def _cancellableInlineCallbacks(
 
         return status.deferred
 
-    _inlineCallbacks(None, gen, status)
+    _inlineCallbacks(None, gen, status, _copy_context())
 
     return deferred
 
