@@ -51,6 +51,7 @@ from twisted.internet.interfaces import (
     IReactorFDSet,
     IReactorSocket,
     IReactorTCP,
+    IReactorTime,
     IResolverSimple,
     ITLSTransport,
 )
@@ -1230,12 +1231,27 @@ def assertPeerClosedOnEMFILE(
 
     serverConnectionMade.addCallback(stopReactorIfServerAccepted)
 
-    port = listen(reactor, serverFactory)
-    listeningHost = port.getHost()
     clientFactory = MyClientFactory()
-    connect(reactor, listeningHost, clientFactory)
 
-    reactor.callWhenRunning(exhauster.exhaust)
+    if IReactorTime.providedBy(reactor):
+        # For Glib-based reactors, the exhauster should be run after the signal handler used by glib [1]
+        # and restoration of twisted signal handlers [2], thus such 2-level callLater
+        # [1] https://gitlab.gnome.org/GNOME/pygobject/-/blob/3.42.0/gi/_ossighelper.py#L76
+        # [2] https://github.com/twisted/twisted/blob/twisted-22.4.0/src/twisted/internet/_glibbase.py#L134
+        # See also https://twistedmatrix.com/trac/ticket/10342
+        def inner():
+            port = listen(reactor, serverFactory)
+            listeningHost = port.getHost()
+            connect(reactor, listeningHost, clientFactory)
+            exhauster.exhaust()
+
+        reactor.callLater(0, reactor.callLater, 0, inner)
+    else:
+        # For reactors without callLater (ex: MemoryReactor)
+        port = listen(reactor, serverFactory)
+        listeningHost = port.getHost()
+        connect(reactor, listeningHost, clientFactory)
+        reactor.callWhenRunning(exhauster.exhaust)
 
     def stopReactorAndCloseFileDescriptors(result):
         exhauster.release()
