@@ -997,10 +997,10 @@ class OpenSSLOptionsTests(OpenSSLOptionsTestsMixin, TestCase):
         self.assertEqual(DeprecationWarning, warnings[0]["category"])
         self.assertEqual(message, warnings[0]["message"])
 
-    def test_tlsv1ByDefault(self):
+    def test_tlsv12ByDefault(self):
         """
         L{sslverify.OpenSSLCertificateOptions} will make the default minimum
-        TLS version v1.0, if no C{method}, or C{insecurelyLowerMinimumTo} is
+        TLS version v1.2, if no C{method}, or C{insecurelyLowerMinimumTo} is
         given.
         """
         opts = sslverify.OpenSSLCertificateOptions(
@@ -1013,6 +1013,8 @@ class OpenSSLOptionsTests(OpenSSLOptionsTestsMixin, TestCase):
             | SSL.OP_NO_COMPRESSION
             | SSL.OP_CIPHER_SERVER_PREFERENCE
             | SSL.OP_NO_SSLv3
+            | SSL.OP_NO_TLSv1
+            | SSL.OP_NO_TLSv1_1
         )
         self.assertEqual(options, ctx._options & options)
 
@@ -1080,7 +1082,7 @@ class OpenSSLOptionsTests(OpenSSLOptionsTestsMixin, TestCase):
             sslverify.OpenSSLCertificateOptions(
                 privateKey=self.sKey,
                 certificate=self.sCert,
-                method=SSL.SSLv23_METHOD,
+                method=SSL.TLS_METHOD,
                 lowerMaximumSecurityTo=sslverify.TLSVersion.TLSv1_2,
             )
 
@@ -1322,9 +1324,11 @@ class OpenSSLOptionsTests(OpenSSLOptionsTestsMixin, TestCase):
             | SSL.OP_NO_COMPRESSION
             | SSL.OP_CIPHER_SERVER_PREFERENCE
             | SSL.OP_NO_SSLv3
+            | SSL.OP_NO_TLSv1
+            | SSL.OP_NO_TLSv1_1
         )
         self.assertEqual(options, ctx._options & options)
-        self.assertEqual(opts._defaultMinimumTLSVersion, sslverify.TLSVersion.TLSv1_0)
+        self.assertEqual(opts._defaultMinimumTLSVersion, sslverify.TLSVersion.TLSv1_2)
 
     def test_tlsProtocolsAllSecureTLS(self):
         """
@@ -3341,31 +3345,19 @@ class SelectVerifyImplementationTests(SynchronousTestCase):
             if warning["category"] == UserWarning
         )
 
-        importErrors = [
-            # Python 3.6.3
-            "'import of service_identity halted; None in sys.modules'",
-            # Python 3
-            "'import of 'service_identity' halted; None in sys.modules'",
-            # Python 2
-            "'No module named service_identity'",
-        ]
+        expectedMessage = (
+            "You do not have a working installation of the "
+            "service_identity module: "
+            "'import of service_identity halted; None in sys.modules'.  "
+            "Please install it from "
+            "<https://pypi.python.org/pypi/service_identity> "
+            "and make sure all of its dependencies are satisfied.  "
+            "Without the service_identity module, Twisted can perform only"
+            " rudimentary TLS client hostname verification.  Many valid "
+            "certificate/hostname mappings may be rejected."
+        )
 
-        expectedMessages = []
-        for importError in importErrors:
-            expectedMessages.append(
-                "You do not have a working installation of the "
-                "service_identity module: {message}.  Please install it from "
-                "<https://pypi.python.org/pypi/service_identity> "
-                "and make sure all of its dependencies are satisfied.  "
-                "Without the service_identity module, Twisted can perform only"
-                " rudimentary TLS client hostname verification.  Many valid "
-                "certificate/hostname mappings may be rejected.".format(
-                    message=importError
-                )
-            )
-
-        self.assertIn(warning["message"], expectedMessages)
-
+        self.assertEqual(warning["message"], expectedMessage)
         # Make sure we're abusing the warning system to a sufficient
         # degree: there is no filename or line number that makes sense for
         # this warning to "blame" for the problem.  It is a system
