@@ -11,99 +11,171 @@ responsible for coordinating all of trial's behavior at the highest level.
 
 import os
 import sys
+from functools import partial
+from typing import Awaitable, Callable, Iterable, List, Sequence, TextIO, Union, cast
+from unittest import TestResult, TestSuite
 
-from twisted.internet.defer import DeferredList
-from twisted.internet.task import cooperate
+from attrs import define, field, frozen
+from attrs.converters import default_if_none
+
+from twisted.internet.defer import Deferred, DeferredList, gatherResults
+from twisted.internet.interfaces import IReactorCore, IReactorProcess
+from twisted.logger import Logger
+from twisted.python.failure import Failure
 from twisted.python.filepath import FilePath
+from twisted.python.lockfile import FilesystemLock
 from twisted.python.modules import theSystemPath
-from twisted.trial._asyncrunner import _iterateTests
-from twisted.trial._dist import _WORKER_AMP_STDIN, _WORKER_AMP_STDOUT
-from twisted.trial._dist.distreporter import DistReporter
-from twisted.trial._dist.worker import LocalWorker, LocalWorkerAMP
-from twisted.trial.reporter import UncleanWarningsReporterWrapper
-from twisted.trial.util import _unusedTestDirectory
+from .._asyncrunner import _iterateTests
+from ..itrial import IReporter, ITestCase
+from ..reporter import UncleanWarningsReporterWrapper
+from ..runner import TestHolder
+from ..util import _unusedTestDirectory, openTestLog
+from . import _WORKER_AMP_STDIN, _WORKER_AMP_STDOUT
+from .distreporter import DistReporter
+from .functional import countingCalls, discardResult, iterateWhile, takeWhile
+from .worker import LocalWorker, LocalWorkerAMP, WorkerAction
 
 
-class DistTrialRunner:
+class IDistTrialReactor(IReactorCore, IReactorProcess):
     """
-    A specialized runner for distributed trial. The runner launches a number of
-    local worker processes which will run tests.
-
-    @ivar _workerNumber: the number of workers to be spawned.
-    @type _workerNumber: C{int}
-
-    @ivar _stream: stream which the reporter will use.
-
-    @ivar _reporterFactory: the reporter class to be used.
+    The reactor interfaces required by disttrial.
     """
 
-    _distReporterFactory = DistReporter
 
-    def _makeResult(self):
-        """
-        Make reporter factory, and wrap it with a L{DistReporter}.
-        """
-        reporter = self._reporterFactory(
-            self._stream, self._tbformat, realtime=self._rterrors
-        )
-        if self._uncleanWarnings:
-            reporter = UncleanWarningsReporterWrapper(reporter)
-        return self._distReporterFactory(reporter)
+def _defaultReactor() -> IDistTrialReactor:
+    """
+    Get the default reactor, ensuring it is suitable for use with disttrial.
+    """
+    import twisted.internet.reactor as defaultReactor
 
-    def __init__(
-        self,
-        reporterFactory,
-        workerNumber,
-        workerArguments,
-        stream=None,
-        tracebackFormat="default",
-        realTimeErrors=False,
-        uncleanWarnings=False,
-        logfile="test.log",
-        workingDirectory="_trial_temp",
+    if all(
+        [
+            IReactorCore.providedBy(defaultReactor),
+            IReactorProcess.providedBy(defaultReactor),
+        ]
     ):
-        self._workerNumber = workerNumber
-        self._workerArguments = workerArguments
-        self._reporterFactory = reporterFactory
-        if stream is None:
-            stream = sys.stdout
-        self._stream = stream
-        self._tbformat = tracebackFormat
-        self._rterrors = realTimeErrors
-        self._uncleanWarnings = uncleanWarnings
-        self._result = None
-        self._workingDirectory = workingDirectory
-        self._logFile = logfile
-        self._logFileObserver = None
-        self._logFileObject = None
-        self._logWarnings = False
+        # If it provides each of the interfaces then it provides the
+        # intersection interface.  cast it to make it easier to talk about
+        # later on.
+        return cast(IDistTrialReactor, defaultReactor)
 
-    def writeResults(self, result):
+    raise TypeError("Reactor does not provide the right interfaces")
+
+
+@frozen
+class WorkerPoolConfig:
+    """
+    Configuration parameters for a pool of test-running workers.
+
+    @ivar numWorkers: The number of workers in the pool.
+
+    @ivar workingDirectory: A directory in which working directories for each
+        of the workers will be created.
+
+    @ivar workerArguments: Extra arguments to pass the worker process in its
+        argv.
+
+    @ivar logFile: The basename of the overall test log file.
+    """
+
+    numWorkers: int
+    workingDirectory: FilePath
+    workerArguments: Sequence[str]
+    logFile: str
+
+
+@define
+class StartedWorkerPool:
+    """
+    A pool of workers which have already been started.
+
+    @ivar workingDirectory: A directory holding the working directories for
+        each of the workers.
+
+    @ivar testDirLock: An object representing the cooperative lock this pool
+        holds on its working directory.
+
+    @ivar testLog: The open overall test log file.
+
+    @ivar workers: Objects corresponding to the worker child processes and
+        adapting between process-related interfaces and C{IProtocol}.
+
+    @ivar ampWorkers: AMP protocol instances corresponding to the worker child
+        processes.
+    """
+
+    workingDirectory: FilePath
+    testDirLock: FilesystemLock
+    testLog: TextIO
+    workers: List[LocalWorker]
+    ampWorkers: List[LocalWorkerAMP]
+
+    _logger = Logger()
+
+    async def run(self, workerAction: WorkerAction) -> None:
         """
-        Write test run final outcome to result.
-
-        @param result: A C{TestResult} which will print errors and the summary.
+        Run an action on all of the workers in the pool.
         """
-        result.done()
+        await gatherResults(
+            discardResult(workerAction(worker)) for worker in self.ampWorkers
+        )
+        return None
 
-    def createLocalWorkers(self, protocols, workingDirectory):
+    async def join(self) -> None:
+        """
+        Shut down all of the workers in the pool.
+
+        The pool is unusable after this method is called.
+        """
+        results = await DeferredList(
+            [Deferred.fromCoroutine(worker.exit()) for worker in self.workers],
+            consumeErrors=True,
+        )
+        for n, (succeeded, failure) in enumerate(results):
+            if not succeeded:
+                self._logger.failure(f"joining disttrial worker #{n} failed", failure)
+
+        del self.workers[:]
+        del self.ampWorkers[:]
+        self.testLog.close()
+        self.testDirLock.unlock()
+
+
+@frozen
+class WorkerPool:
+    """
+    Manage a fixed-size collection of child processes which can run tests.
+
+    @ivar _config: Configuration for the precise way in which the pool is run.
+    """
+
+    _config: WorkerPoolConfig
+
+    def _createLocalWorkers(
+        self,
+        protocols: Iterable[LocalWorkerAMP],
+        workingDirectory: FilePath,
+        logFile: TextIO,
+    ) -> List[LocalWorker]:
         """
         Create local worker protocol instances and return them.
 
-        @param protocols: An iterable of L{LocalWorkerAMP} instances.
+        @param protocols: The process/protocol adapters to use for the created
+        workers.
 
         @param workingDirectory: The base path in which we should run the
             workers.
-        @type workingDirectory: C{str}
+
+        @param logFile: The test log, for workers to write to.
 
         @return: A list of C{quantity} C{LocalWorker} instances.
         """
         return [
-            LocalWorker(protocol, os.path.join(workingDirectory, str(x)), self._logFile)
+            LocalWorker(protocol, workingDirectory.child(str(x)), logFile)
             for x, protocol in enumerate(protocols)
         ]
 
-    def launchWorkerProcesses(self, spawner, protocols, arguments):
+    def _launchWorkerProcesses(self, spawner, protocols, arguments):
         """
         Spawn processes from a list of process protocols.
 
@@ -130,7 +202,137 @@ class DistTrialRunner:
             args.extend(arguments)
             spawner(worker, sys.executable, args=args, childFDs=childFDs, env=environ)
 
-    def _driveWorker(self, worker, result, testCases, cooperate):
+    async def start(self, reactor: IReactorProcess) -> StartedWorkerPool:
+        """
+        Launch all of the workers for this pool.
+
+        @return: A started pool object that can run jobs using the workers.
+        """
+        testDir, testDirLock = _unusedTestDirectory(
+            self._config.workingDirectory,
+        )
+
+        # Open a log file in the chosen working directory (not necessarily the
+        # same as our configured working directory, if that path was in use).
+        testLog = openTestLog(testDir.child(self._config.logFile))
+
+        ampWorkers = [LocalWorkerAMP() for x in range(self._config.numWorkers)]
+        workers = self._createLocalWorkers(
+            ampWorkers,
+            testDir,
+            testLog,
+        )
+        self._launchWorkerProcesses(
+            reactor.spawnProcess,
+            workers,
+            self._config.workerArguments,
+        )
+
+        return StartedWorkerPool(
+            testDir,
+            testDirLock,
+            testLog,
+            workers,
+            ampWorkers,
+        )
+
+
+def shouldContinue(untilFailure: bool, result: IReporter) -> bool:
+    """
+    Determine whether the test suite should be iterated again.
+
+    @param untilFailure: C{True} if the suite is supposed to run until
+        failure.
+
+    @param result: The test result of the test suite iteration which just
+        completed.
+    """
+    return untilFailure and result.wasSuccessful()
+
+
+async def runTests(
+    pool: StartedWorkerPool,
+    testCases: Iterable[ITestCase],
+    result: DistReporter,
+    driveWorker: Callable[
+        [DistReporter, Sequence[ITestCase], LocalWorkerAMP], Awaitable[None]
+    ],
+) -> None:
+    try:
+        # Run the tests using the worker pool.
+        await pool.run(partial(driveWorker, result, testCases))
+    except Exception:
+        # Exceptions from test code are handled somewhere else.  An
+        # exception here is a bug in the runner itself.  The only
+        # convenient place to put it is in the result, though.
+        result.original.addError(TestHolder("<runTests>"), Failure())
+
+
+@define
+class DistTrialRunner:
+    """
+    A specialized runner for distributed trial. The runner launches a number of
+    local worker processes which will run tests.
+
+    @ivar _maxWorkers: the number of workers to be spawned.
+
+    @ivar _exitFirst: ``True`` to stop the run as soon as a test case fails.
+        ``False`` to run through the whole suite and report all of the results
+        at the end.
+
+    @ivar _stream: stream which the reporter will use.
+
+    @ivar _reporterFactory: the reporter class to be used.
+    """
+
+    _distReporterFactory = DistReporter
+    _logger = Logger()
+
+    # accepts a `realtime` keyword argument which we can't annotate, so punt
+    # on the argument annotation
+    _reporterFactory: Callable[..., IReporter]
+    _maxWorkers: int
+    _workerArguments: List[str]
+    _exitFirst: bool = False
+    _reactor: IDistTrialReactor = field(
+        # mypy doesn't understand the converter
+        default=None,
+        converter=default_if_none(factory=_defaultReactor),  # type: ignore [misc]
+    )
+    # mypy doesn't understand the converter
+    _stream: TextIO = field(default=None, converter=default_if_none(sys.stdout))  # type: ignore [misc]
+    _tracebackFormat: str = "default"
+    _realTimeErrors: bool = False
+    _uncleanWarnings: bool = False
+    _logfile: str = "test.log"
+    _workingDirectory: str = "_trial_temp"
+    _workerPoolFactory: Callable[[WorkerPoolConfig], WorkerPool] = WorkerPool
+
+    def _makeResult(self) -> DistReporter:
+        """
+        Make reporter factory, and wrap it with a L{DistReporter}.
+        """
+        reporter = self._reporterFactory(
+            self._stream, self._tracebackFormat, realtime=self._realTimeErrors
+        )
+        if self._uncleanWarnings:
+            reporter = UncleanWarningsReporterWrapper(reporter)
+        return self._distReporterFactory(reporter)
+
+    def writeResults(self, result):
+        """
+        Write test run final outcome to result.
+
+        @param result: A C{TestResult} which will print errors and the summary.
+        """
+        result.done()
+
+    async def _driveWorker(
+        self,
+        result: DistReporter,
+        testCases: Sequence[ITestCase],
+        worker: LocalWorkerAMP,
+    ) -> None:
         """
         Drive a L{LocalWorkerAMP} instance, iterating the tests and calling
         C{run} for every one of them.
@@ -141,113 +343,123 @@ class DistTrialRunner:
 
         @param testCases: The global list of tests to iterate.
 
-        @param cooperate: The cooperate function to use, to be customized in
-            tests.
-        @type cooperate: C{function}
-
-        @return: A C{Deferred} firing when all the tests are finished.
+        @return: A coroutine that completes after all of the tests have
+            completed.
         """
 
-        def resultErrback(error, case):
-            result.original.addFailure(case, error)
-            return error
+        async def task(case):
+            try:
+                await worker.run(case, result)
+            except Exception:
+                result.original.addError(case, Failure())
 
-        def task(case):
-            d = worker.run(case, result)
-            d.addErrback(resultErrback, case)
-            return d
+        for case in testCases:
+            await task(case)
 
-        return cooperate(task(case) for case in testCases).whenDone()
-
-    def run(self, suite, reactor=None, cooperate=cooperate, untilFailure=False):
+    async def runAsync(
+        self,
+        suite: TestSuite,
+        untilFailure: bool = False,
+    ) -> DistReporter:
         """
         Spawn local worker processes and load tests. After that, run them.
 
         @param suite: A tests suite to be run.
 
-        @param reactor: The reactor to use, to be customized in tests.
-        @type reactor: A provider of
-            L{twisted.internet.interfaces.IReactorProcess}
-
-        @param cooperate: The cooperate function to use, to be customized in
-            tests.
-        @type cooperate: C{function}
-
         @param untilFailure: If C{True}, continue to run the tests until they
             fail.
-        @type untilFailure: C{bool}.
 
-        @return: The test result.
-        @rtype: L{DistReporter}
+        @return: A coroutine that completes with the test result.
         """
-        if reactor is None:
-            from twisted.internet import reactor
-        result = self._makeResult()
-        count = suite.countTestCases()
-        self._stream.write("Running %d tests.\n" % (count,))
 
-        if not count:
-            # Take a shortcut if there is no test
-            suite.run(result.original)
+        # Realize a concrete set of tests to run.
+        testCases = list(_iterateTests(suite))
+
+        # Create a worker pool to use to execute them.
+        poolStarter = self._workerPoolFactory(
+            WorkerPoolConfig(
+                # Don't make it larger than is useful or allowed.
+                min(len(testCases), self._maxWorkers),
+                FilePath(self._workingDirectory),
+                self._workerArguments,
+                self._logfile,
+            ),
+        )
+
+        # Announce that we're beginning.  countTestCases result is preferred
+        # (over len(testCases)) because testCases may contain synthetic cases
+        # for error reporting purposes.
+        self._stream.write(f"Running {suite.countTestCases()} tests.\n")
+
+        # Start the worker pool.
+        startedPool = await poolStarter.start(self._reactor)
+
+        # The condition that will determine whether the test run repeats.
+        condition = partial(shouldContinue, untilFailure)
+
+        # A function that will run the whole suite once.
+        @countingCalls
+        async def runAndReport(n: int) -> DistReporter:
+            if untilFailure:
+                # If and only if we're running the suite more than once,
+                # provide a report about which run this is.
+                self._stream.write(f"Test Pass {n + 1}\n")
+
+            result = self._makeResult()
+
+            if self._exitFirst:
+                # Keep giving out tests as long as the result object has only
+                # seen success.
+                casesCondition = lambda _: result.original.wasSuccessful()
+            else:
+                casesCondition = lambda _: True
+
+            await runTests(
+                startedPool,
+                takeWhile(casesCondition, testCases),
+                result,
+                self._driveWorker,
+            )
             self.writeResults(result)
             return result
 
-        testDir, testDirLock = _unusedTestDirectory(FilePath(self._workingDirectory))
-        workerNumber = min(count, self._workerNumber)
-        ampWorkers = [LocalWorkerAMP() for x in range(workerNumber)]
-        workers = self.createLocalWorkers(ampWorkers, testDir.path)
-        processEndDeferreds = [worker.endDeferred for worker in workers]
-        self.launchWorkerProcesses(reactor.spawnProcess, workers, self._workerArguments)
+        try:
+            # Start submitting tests to workers in the pool.  Perhaps repeat
+            # the whole test suite more than once, if appropriate for our
+            # configuration.
+            return await iterateWhile(condition, runAndReport)
+        finally:
+            # Shut down the worker pool.
+            await startedPool.join()
 
-        def runTests():
-            testCases = iter(list(_iterateTests(suite)))
+    def run(self, suite: TestSuite, untilFailure: bool = False) -> TestResult:
+        """
+        Run a reactor and a test suite.
 
-            workerDeferreds = []
-            for worker in ampWorkers:
-                workerDeferreds.append(
-                    self._driveWorker(worker, result, testCases, cooperate=cooperate)
-                )
-            return DeferredList(
-                workerDeferreds, consumeErrors=True, fireOnOneErrback=True
-            )
+        @param suite: The test suite to run.
+        """
+        result: Union[Failure, DistReporter]
 
-        stopping = []
+        def capture(r):
+            nonlocal result
+            result = r
 
-        def nextRun(ign, n):
-            if untilFailure:
-                self._stream.write("Test Pass %d\n" % (n,))
-            self.writeResults(result)
-            if not untilFailure:
-                return
-            if not result.wasSuccessful():
-                return
-            d = runTests()
-            return d.addCallback(nextRun, n + 1)
+        d = Deferred.fromCoroutine(self.runAsync(suite, untilFailure))
+        d.addBoth(capture)
+        d.addBoth(lambda ignored: self._reactor.stop())
+        self._reactor.run()
 
-        def stop(ign):
-            testDirLock.unlock()
-            if not stopping:
-                stopping.append(None)
-                reactor.stop()
+        if isinstance(result, Failure):
+            result.raiseException()
 
-        def beforeShutDown():
-            if not stopping:
-                stopping.append(None)
-                d = DeferredList(processEndDeferreds, consumeErrors=True)
-                return d.addCallback(continueShutdown)
+        # mypy can't see that raiseException raises an exception so we can
+        # only get here if result is not a Failure, so tell mypy result is
+        # certainly a DistReporter at this point.
+        assert isinstance(result, DistReporter)
 
-        def continueShutdown(ign):
-            self.writeResults(result)
-            return ign
-
-        d = runTests()
-        d.addCallback(nextRun, 1)
-        d.addBoth(stop)
-
-        reactor.addSystemEventTrigger("before", "shutdown", beforeShutDown)
-        reactor.run()
-
-        return result
+        # Unwrap the DistReporter to give the caller some regular TestResult
+        # object.  DistReporter isn't type annotated correctly so fix it here.
+        return cast(TestResult, result.original)
 
     def runUntilFailure(self, suite):
         """
