@@ -7,40 +7,66 @@ Tests for L{twisted.trial._dist.disttrial}.
 
 import os
 import sys
+from functools import partial
 from io import StringIO
+from typing import Callable, List, Set
 
 from zope.interface import implementer, verify
 
-from twisted.internet import error, interfaces, reactor
-from twisted.internet.defer import fail, gatherResults, maybeDeferred, succeed
-from twisted.internet.main import CONNECTION_DONE
+from attrs import Factory, define, field
+from hamcrest import assert_that, contains, equal_to, has_length, none
+
+from twisted.internet import interfaces
+from twisted.internet.base import ReactorBase
+from twisted.internet.defer import Deferred, succeed
+from twisted.internet.error import ProcessDone
 from twisted.internet.protocol import ProcessProtocol, Protocol
-from twisted.internet.task import Cooperator, deferLater
+from twisted.internet.test.modulehelpers import AlternateReactor
 from twisted.python.failure import Failure
+from twisted.python.filepath import FilePath
 from twisted.python.lockfile import FilesystemLock
 from twisted.test.proto_helpers import MemoryReactorClock
-from twisted.test.test_cooperator import FakeScheduler
+from twisted.trial._dist import _WORKER_AMP_STDIN
 from twisted.trial._dist.distreporter import DistReporter
-from twisted.trial._dist.disttrial import DistTrialRunner
-from twisted.trial._dist.worker import LocalWorker
+from twisted.trial._dist.disttrial import DistTrialRunner, WorkerPool, WorkerPoolConfig
+from twisted.trial._dist.functional import (
+    countingCalls,
+    discardResult,
+    fromOptional,
+    iterateWhile,
+    sequence,
+)
+from twisted.trial._dist.worker import LocalWorker, Worker, WorkerAction
 from twisted.trial.reporter import (
     Reporter,
+    TestResult,
     TreeReporter,
     UncleanWarningsReporterWrapper,
 )
 from twisted.trial.runner import ErrorHolder, TrialSuite
 from twisted.trial.unittest import SynchronousTestCase, TestCase
+from ...test import erroneous, sample
+from .matchers import matches_result
 
 
+@define
 class FakeTransport:
     """
     A simple fake process transport.
     """
 
+    _closed: Set[int] = field(default=Factory(set))
+
     def writeToChild(self, fd, data):
         """
         Ignore write calls.
         """
+
+    def closeChildFD(self, fd):
+        """
+        Mark one of the child descriptors as closed.
+        """
+        self._closed.add(fd)
 
 
 @implementer(interfaces.IReactorProcess)
@@ -158,85 +184,36 @@ class CountingReactorTests(SynchronousTestCase):
             self.assertEqual(len(whenRunningCalls), count)
 
 
-class EternalTerminationPredicateFactory:
+class WorkerPoolTests(TestCase):
     """
-    A rigged terminationPredicateFactory for which time never pass.
-    """
-
-    def __call__(self):
-        """
-        See: L{task._Timer}
-        """
-        return False
-
-
-class DistTrialRunnerTests(TestCase):
-    """
-    Tests for L{DistTrialRunner}.
+    Tests for L{WorkerPool}.
     """
 
     def setUp(self):
-        """
-        Create a runner for testing.
-        """
-        self.runner = DistTrialRunner(
-            TreeReporter, 4, [], workingDirectory=self.mktemp()
+        self.parent = FilePath(self.mktemp())
+        self.workingDirectory = self.parent.child("_trial_temp")
+        self.config = WorkerPoolConfig(
+            numWorkers=4,
+            workingDirectory=self.workingDirectory,
+            workerArguments=[],
+            logFile="out.log",
         )
-        self.runner._stream = StringIO()
-
-    def reap(self, workers):
-        """
-        Reap the workers and trap L{ConnectionDone} failures on their
-        C{endDeferred}s.
-
-        @param workers: The workers to reap.
-        @type workers: An iterable of L{LocalWorker}
-        """
-
-        for worker in workers:
-            worker.endDeferred.addErrback(Failure.trap, error.ConnectionDone)
-            worker.processEnded(Failure(CONNECTION_DONE))
-
-    def getFakeSchedulerAndEternalCooperator(self):
-        """
-        Helper to create fake scheduler and cooperator in tests.
-
-        The cooperator has a termination timer which will never inform
-        the scheduler that the task needs to be terminated.
-
-        @return: L{tuple} of (scheduler, cooperator)
-        """
-        scheduler = FakeScheduler()
-        cooperator = Cooperator(
-            scheduler=scheduler,
-            terminationPredicateFactory=EternalTerminationPredicateFactory,
-        )
-        return scheduler, cooperator
-
-    def test_writeResults(self):
-        """
-        L{DistTrialRunner.writeResults} writes to the stream specified in the
-        init.
-        """
-        stringIO = StringIO()
-        result = DistReporter(Reporter(stringIO))
-        self.runner.writeResults(result)
-        self.assertTrue(stringIO.tell() > 0)
+        self.pool = WorkerPool(self.config)
 
     def test_createLocalWorkers(self):
         """
-        C{createLocalWorkers} iterates the list of protocols and create one
+        C{_createLocalWorkers} iterates the list of protocols and create one
         L{LocalWorker} for each.
         """
         protocols = [object() for x in range(4)]
-        workers = self.runner.createLocalWorkers(protocols, "path")
+        workers = self.pool._createLocalWorkers(protocols, FilePath("path"), StringIO())
         for s in workers:
             self.assertIsInstance(s, LocalWorker)
         self.assertEqual(4, len(workers))
 
     def test_launchWorkerProcesses(self):
         """
-        Given a C{spawnProcess} function, C{launchWorkerProcess} launches a
+        Given a C{spawnProcess} function, C{_launchWorkerProcess} launches a
         python process with an existing path as its argument.
         """
         protocols = [ProcessProtocol() for i in range(4)]
@@ -258,7 +235,7 @@ class DistTrialRunnerTests(TestCase):
             arguments.extend(args)
             environment.update(env)
 
-        self.runner.launchWorkerProcesses(fakeSpawnProcess, protocols, ["foo"])
+        self.pool._launchWorkerProcesses(fakeSpawnProcess, protocols, ["foo"])
         self.assertEqual(arguments[0], arguments[1])
         self.assertTrue(os.path.exists(arguments[2]))
         self.assertEqual("foo", arguments[3])
@@ -269,97 +246,141 @@ class DistTrialRunnerTests(TestCase):
 
     def test_run(self):
         """
-        C{run} starts the reactor exactly once and spawns each of the workers
-        exactly once.
+        C{run} dispatches the given action to each of its workers exactly once.
         """
-        workers = []
-        fakeReactor = CountingReactor(workers)
-        self.addCleanup(self.reap, workers)
+        # Make sure the parent of the working directory exists so
+        # manage a lock in it.
+        self.parent.makedirs()
 
-        suite = TrialSuite()
-        for i in range(10):
-            suite.addTest(TestCase())
-        self.runner.run(suite, fakeReactor)
-        self.assertEqual(fakeReactor.runCount, 1)
-        self.assertEqual(fakeReactor.spawnCount, self.runner._workerNumber)
+        workers = []
+        starting = self.pool.start(CountingReactor([]))
+        started = self.successResultOf(starting)
+        running = started.run(lambda w: succeed(workers.append(w)))
+        self.successResultOf(running)
+        assert_that(workers, has_length(self.config.numWorkers))
 
     def test_runUsedDirectory(self):
         """
-        L{DistTrialRunner} checks if the test directory is already locked, and
-        if it is generates a name based on it.
+        L{WorkerPool.start} checks if the test directory is already locked, and if
+        it is generates a name based on it.
         """
+        # Make sure the parent of the working directory exists so we can
+        # manage a lock in it.
+        self.parent.makedirs()
 
-        class CountingReactorWithLock(CountingReactor):
-            def spawnProcess(oself, worker, *args, **kwargs):
-                oself._workers.append(worker)
-                self.assertEqual(
-                    os.path.abspath(worker._logDirectory),
-                    os.path.abspath(
-                        os.path.join(workingDirectory + "-1", str(oself.spawnCount))
-                    ),
-                )
-                localLock = FilesystemLock(workingDirectory + "-1.lock")
-                self.assertFalse(localLock.lock())
-                oself.spawnCount += 1
-                worker.makeConnection(FakeTransport())
-                worker._ampProtocol.run = lambda *args: succeed(None)
-
-        newDirectory = self.mktemp()
-        os.mkdir(newDirectory)
-        workingDirectory = os.path.join(newDirectory, "_trial_temp")
-        lock = FilesystemLock(workingDirectory + ".lock")
-        lock.lock()
+        # Lock the directory the runner will expect to use.
+        lock = FilesystemLock(self.workingDirectory.path + ".lock")
+        self.assertTrue(lock.lock())
         self.addCleanup(lock.unlock)
-        self.runner._workingDirectory = workingDirectory
 
-        workers = []
+        # Start up the pool
+        fakeReactor = CountingReactor([])
+        started = self.successResultOf(self.pool.start(fakeReactor))
 
-        fakeReactor = CountingReactorWithLock(workers)
-        self.addCleanup(self.reap, workers)
+        # Verify it took a nearby directory instead.
+        self.assertEqual(
+            started.workingDirectory,
+            self.workingDirectory.sibling("_trial_temp-1"),
+        )
 
-        suite = TrialSuite()
-        for i in range(10):
-            suite.addTest(TestCase())
+    def test_join(self):
+        """
+        L{StartedWorkerPool.join} causes all of the workers to exit, closes the
+        log file, and unlocks the test directory.
+        """
+        self.parent.makedirs()
 
-        self.runner.run(suite, fakeReactor)
+        reactor = CountingReactor([])
+        started = self.successResultOf(self.pool.start(reactor))
+        joining = Deferred.fromCoroutine(started.join())
+        self.assertNoResult(joining)
+        for w in reactor._workers:
+            assert_that(w.transport._closed, contains(_WORKER_AMP_STDIN))
+            for fd in w.transport._closed:
+                w.childConnectionLost(fd)
+            for f in [w.processExited, w.processEnded]:
+                f(Failure(ProcessDone(0)))
+        assert_that(self.successResultOf(joining), none())
+        assert_that(started.testLog.closed, equal_to(True))
+        assert_that(started.testDirLock.locked, equal_to(False))
+
+
+class DistTrialRunnerTests(TestCase):
+    """
+    Tests for L{DistTrialRunner}.
+    """
+
+    suite = TrialSuite([sample.FooTest("test_foo")])
+
+    def getRunner(self, **overrides):
+        """
+        Create a runner for testing.
+        """
+        args = dict(
+            reporterFactory=TreeReporter,
+            workingDirectory=self.mktemp(),
+            stream=StringIO(),
+            maxWorkers=4,
+            workerArguments=[],
+            workerPoolFactory=partial(LocalWorkerPool, autostop=True),
+            reactor=CountingReactor([]),
+        )
+        args.update(overrides)
+        return DistTrialRunner(**args)
+
+    def test_writeResults(self):
+        """
+        L{DistTrialRunner.writeResults} writes to the stream specified in the
+        init.
+        """
+        stringIO = StringIO()
+        result = DistReporter(Reporter(stringIO))
+        runner = self.getRunner()
+        runner.writeResults(result)
+        self.assertTrue(stringIO.tell() > 0)
 
     def test_minimalWorker(self):
         """
-        L{DistTrialRunner} doesn't try to start more workers than the number of
-        tests.
+        L{DistTrialRunner.runAsync} doesn't try to start more workers than the
+        number of tests.
         """
-        workers = []
-        fakeReactor = CountingReactor(workers)
-        self.addCleanup(self.reap, workers)
+        pool = None
 
-        self.runner.run(TestCase(), fakeReactor)
-        self.assertEqual(fakeReactor.runCount, 1)
-        self.assertEqual(fakeReactor.spawnCount, 1)
+        def recordingFactory(*a, **kw):
+            nonlocal pool
+            pool = LocalWorkerPool(*a, autostop=True, **kw)
+            return pool
 
-    def test_runUncleanWarnings(self):
+        maxWorkers = 7
+        numTests = 3
+
+        runner = self.getRunner(
+            maxWorkers=maxWorkers, workerPoolFactory=recordingFactory
+        )
+        suite = TrialSuite([TestCase() for n in range(numTests)])
+        self.successResultOf(runner.runAsync(suite))
+        assert_that(pool._started[0].workers, has_length(numTests))
+
+    def test_runUncleanWarnings(self) -> None:
         """
-        Running with the C{unclean-warnings} option makes L{DistTrialRunner}
-        uses the L{UncleanWarningsReporterWrapper}.
+        Running with the C{unclean-warnings} option makes L{DistTrialRunner} uses
+        the L{UncleanWarningsReporterWrapper}.
         """
-        workers = []
-        fakeReactor = CountingReactor(workers)
-        self.addCleanup(self.reap, workers)
-
-        self.runner._uncleanWarnings = True
-        result = self.runner.run(TestCase(), fakeReactor)
+        runner = self.getRunner(uncleanWarnings=True)
+        d = runner.runAsync(self.suite)
+        result = self.successResultOf(d)
         self.assertIsInstance(result, DistReporter)
         self.assertIsInstance(result.original, UncleanWarningsReporterWrapper)
 
     def test_runWithoutTest(self):
         """
-        When the suite contains no test, L{DistTrialRunner} takes a shortcut
-        path without launching any process or starting the reactor.
+        L{DistTrialRunner} can run an empty test suite.
         """
-        fakeReactor = object()
-        suite = TrialSuite()
-        result = self.runner.run(suite, fakeReactor)
+        stream = StringIO()
+        runner = self.getRunner(stream=stream)
+        result = self.successResultOf(runner.runAsync(TrialSuite()))
         self.assertIsInstance(result, DistReporter)
-        output = self.runner._stream.getvalue()
+        output = stream.getvalue()
         self.assertIn("Running 0 test", output)
         self.assertIn("PASSED", output)
 
@@ -369,146 +390,406 @@ class DistTrialRunnerTests(TestCase):
         an import error): this should make the run fail, and the error should
         be printed.
         """
-        fakeReactor = object()
-        error = ErrorHolder("an error", Failure(RuntimeError("foo bar")))
-        result = self.runner.run(error, fakeReactor)
+        err = ErrorHolder("an error", Failure(RuntimeError("foo bar")))
+        stream = StringIO()
+        runner = self.getRunner(stream=stream)
+
+        result = self.successResultOf(runner.runAsync(err))
         self.assertIsInstance(result, DistReporter)
-        output = self.runner._stream.getvalue()
+        output = stream.getvalue()
         self.assertIn("Running 0 test", output)
         self.assertIn("foo bar", output)
         self.assertIn("an error", output)
         self.assertIn("errors=1", output)
         self.assertIn("FAILED", output)
 
-    def test_runUnexpectedError(self):
+    def test_runUnexpectedError(self) -> None:
         """
-        If for some reasons we can't connect to the worker process, the test
-        suite catches and fails.
+        If for some reasons we can't connect to the worker process, the error is
+        recorded in the result object.
         """
+        runner = self.getRunner(workerPoolFactory=BrokenWorkerPool)
+        result = self.successResultOf(runner.runAsync(self.suite))
+        errors = result.original.errors
+        assert_that(errors, has_length(1))
+        assert_that(errors[0][1].type, equal_to(WorkerPoolBroken))
 
-        class CountingReactorWithFail(CountingReactor):
-            def spawnProcess(self, worker, *args, **kwargs):
-                self._workers.append(worker)
-                worker.makeConnection(FakeTransport())
-                self.spawnCount += 1
-                worker._ampProtocol.run = self.failingRun
-
-            def failingRun(self, case, result):
-                return fail(RuntimeError("oops"))
-
-        scheduler, cooperator = self.getFakeSchedulerAndEternalCooperator()
-
-        workers = []
-        fakeReactor = CountingReactorWithFail(workers)
-        self.addCleanup(self.reap, workers)
-
-        result = self.runner.run(TestCase(), fakeReactor, cooperator.cooperate)
-        self.assertEqual(fakeReactor.runCount, 1)
-        self.assertEqual(fakeReactor.spawnCount, 1)
-        scheduler.pump()
-        self.assertEqual(1, len(result.original.failures))
-
-    def test_runStopAfterTests(self):
+    def test_runUnexpectedWorkerError(self) -> None:
         """
-        L{DistTrialRunner} calls C{reactor.stop} and unlocks the test directory
-        once the tests have run.
+        If for some reason the worker process cannot run a test, the error is
+        recorded in the result object.
         """
+        runner = self.getRunner(
+            workerPoolFactory=partial(
+                LocalWorkerPool, workerFactory=_BrokenLocalWorker, autostop=True
+            )
+        )
+        result = self.successResultOf(runner.runAsync(self.suite))
+        errors = result.original.errors
+        assert_that(errors, has_length(1))
+        assert_that(errors[0][1].type, equal_to(WorkerBroken))
 
-        class CountingReactorWithSuccess(CountingReactor):
-            def spawnProcess(self, worker, *args, **kwargs):
-                self._workers.append(worker)
-                worker.makeConnection(FakeTransport())
-                self.spawnCount += 1
-                worker._ampProtocol.run = self.succeedingRun
-
-            def succeedingRun(self, case, result):
-                return succeed(None)
-
-        workingDirectory = self.runner._workingDirectory
-
-        workers = []
-        fakeReactor = CountingReactorWithSuccess(workers)
-
-        self.runner.run(TestCase(), fakeReactor)
-
-        def check():
-            localLock = FilesystemLock(workingDirectory + ".lock")
-            self.assertTrue(localLock.lock())
-            self.assertEqual(1, fakeReactor.stopCount)
-
-        self.assertEqual(list(fakeReactor.triggers.keys()), ["before"])
-        self.assertEqual(list(fakeReactor.triggers["before"]), ["shutdown"])
-        self.reap(workers)
-
-        return deferLater(reactor, 0, check)
-
-    def test_runWaitForProcessesDeferreds(self):
+    def test_runWaitForProcessesDeferreds(self) -> None:
         """
-        L{DistTrialRunner} waits for the worker processes to stop when the
-        reactor is stopping, and then unlocks the test directory, not trying to
-        stop the reactor again.
+        L{DistTrialRunner} waits for the worker pool to stop.
         """
-        workers = []
-        workingDirectory = self.runner._workingDirectory
+        pool = None
 
-        fakeReactor = CountingReactor(workers)
-        self.runner.run(TestCase(), fakeReactor)
+        def recordingFactory(*a, **kw):
+            nonlocal pool
+            pool = LocalWorkerPool(*a, autostop=False, **kw)
+            return pool
 
-        def check(ign):
-            # Let the AMP deferreds fire
-            return deferLater(reactor, 0, realCheck)
+        runner = self.getRunner(
+            workerPoolFactory=recordingFactory,
+        )
+        d = Deferred.fromCoroutine(runner.runAsync(self.suite))
+        if pool is None:
+            self.fail("worker pool was never created")
 
-        def realCheck():
-            localLock = FilesystemLock(workingDirectory + ".lock")
-            self.assertTrue(localLock.lock())
-            # Stop is not called, as it ought to have been called before
-            self.assertEqual(0, fakeReactor.stopCount)
+        assert pool is not None
+        stopped = pool._started[0]._stopped
+        self.assertNoResult(d)
+        stopped.callback(None)
+        result = self.successResultOf(d)
+        self.assertIsInstance(result, DistReporter)
 
-        self.assertEqual(list(fakeReactor.triggers.keys()), ["before"])
-        self.assertEqual(list(fakeReactor.triggers["before"]), ["shutdown"])
-        self.reap(workers)
-
-        return gatherResults(
+    def test_exitFirst(self):
+        """
+        L{DistTrialRunner} can run in C{exitFirst} mode where it will run until a
+        test fails and then abandon the rest of the suite.
+        """
+        stream = StringIO()
+        # Construct a suite with a failing test in the middle.
+        suite = TrialSuite(
             [
-                maybeDeferred(f, *a, **kw)
-                for f, a, kw in fakeReactor.triggers["before"]["shutdown"]
+                sample.FooTest("test_foo"),
+                erroneous.TestRegularFail("test_fail"),
+                sample.FooTest("test_bar"),
             ]
-        ).addCallback(check)
+        )
+        runner = self.getRunner(stream=stream, exitFirst=True, maxWorkers=2)
+        d = runner.runAsync(suite)
+        result = self.successResultOf(d)
+        assert_that(
+            result.original,
+            matches_result(
+                successes=1,
+                failures=has_length(1),
+            ),
+        )
 
     def test_runUntilFailure(self):
         """
         L{DistTrialRunner} can run in C{untilFailure} mode where it will run
         the given tests until they fail.
         """
-        called = []
-
-        class CountingReactorWithSuccess(CountingReactor):
-            def spawnProcess(self, worker, *args, **kwargs):
-                self._workers.append(worker)
-                worker.makeConnection(FakeTransport())
-                self.spawnCount += 1
-                worker._ampProtocol.run = self.succeedingRun
-
-            def succeedingRun(self, case, result):
-                called.append(None)
-                if len(called) == 5:
-                    return fail(RuntimeError("oops"))
-                return succeed(None)
-
-        workers = []
-        fakeReactor = CountingReactorWithSuccess(workers)
-        self.addCleanup(self.reap, workers)
-
-        scheduler, cooperator = self.getFakeSchedulerAndEternalCooperator()
-
-        result = self.runner.run(
-            TestCase(), fakeReactor, cooperate=cooperator.cooperate, untilFailure=True
-        )
-        scheduler.pump()
-        self.assertEqual(5, len(called))
+        stream = StringIO()
+        case = erroneous.EventuallyFailingTestCase("test_it")
+        runner = self.getRunner(stream=stream)
+        d = runner.runAsync(case, untilFailure=True)
+        result = self.successResultOf(d)
+        # The case is hard-coded to fail on its 5th run.
+        self.assertEqual(5, case.n)
         self.assertFalse(result.wasSuccessful())
-        output = self.runner._stream.getvalue()
-        self.assertIn("PASSED", output)
+        output = stream.getvalue()
+
+        # It passes each time except the last.
+        self.assertEqual(
+            output.count("PASSED"),
+            case.n - 1,
+            "expected to see PASSED in output",
+        )
+        # It also fails at the end.
         self.assertIn("FAIL", output)
-        for i in range(1, 5):
+
+        # It also reports its progress.
+        for i in range(1, 6):
             self.assertIn(f"Test Pass {i}", output)
+
+        # It also reports the number of tests run as part of each iteration.
+        self.assertEqual(
+            output.count("Ran 1 tests in"),
+            case.n,
+            "expected to see per-iteration test count in output",
+        )
+
+    def test_run(self) -> None:
+        """
+        L{DistTrialRunner.run} returns a L{DistReporter} containing the result of
+        the test suite run.
+        """
+        runner = self.getRunner()
+        result = runner.run(self.suite)
+        assert_that(result.wasSuccessful(), equal_to(True))
+        assert_that(result.successes, equal_to(1))
+
+    def test_installedReactor(self) -> None:
+        """
+        L{DistTrialRunner.run} uses the installed reactor L{DistTrialRunner} was
+        constructed without a reactor.
+        """
+        reactor = CountingReactor([])
+        with AlternateReactor(reactor):
+            runner = self.getRunner(reactor=None)
+        result = runner.run(self.suite)
+        assert_that(result.errors, equal_to([]))
+        assert_that(result.failures, equal_to([]))
+        assert_that(result.wasSuccessful(), equal_to(True))
+        assert_that(result.successes, equal_to(1))
+        assert_that(reactor.runCount, equal_to(1))
+        assert_that(reactor.stopCount, equal_to(1))
+
+    def test_wrongInstalledReactor(self) -> None:
+        """
+        L{DistTrialRunner} raises L{TypeError} if the installed reactor provides
+        neither L{IReactorCore} nor L{IReactorProcess} and no other reactor is
+        given.
+        """
+
+        class Core(ReactorBase):
+            def installWaker(self):
+                pass
+
+        @implementer(interfaces.IReactorProcess)
+        class Process:
+            def spawnProcess(
+                self,
+                processProtocol,
+                executable,
+                args,
+                env,
+                path,
+                uid,
+                gid,
+                usePTY,
+                childFDs,
+            ):
+                pass
+
+        class Neither:
+            pass
+
+        # It provides neither
+        with AlternateReactor(Neither()):
+            with self.assertRaises(TypeError):
+                self.getRunner(reactor=None)
+
+        # It is missing IReactorProcess
+        with AlternateReactor(Core()):
+            with self.assertRaises(TypeError):
+                self.getRunner(reactor=None)
+
+        # It is missing IReactorCore
+        with AlternateReactor(Process()):
+            with self.assertRaises(TypeError):
+                self.getRunner(reactor=None)
+
+    def test_runFailure(self):
+        """
+        If there is an unexpected exception running the test suite then it is
+        re-raised by L{DistTrialRunner.run}.
+        """
+        # Give it a broken worker pool factory.  There's no exception handling
+        # for such an error in the implementation..
+        class BrokenFactory(Exception):
+            pass
+
+        def brokenFactory(*args, **kwargs):
+            raise BrokenFactory()
+
+        runner = self.getRunner(workerPoolFactory=brokenFactory)
+        with self.assertRaises(BrokenFactory):
+            runner.run(self.suite)
+
+
+class FunctionalTests(TestCase):
+    """
+    Tests for the functional helpers that need it.
+    """
+
+    def test_fromOptional(self) -> None:
+        """
+        ``fromOptional`` accepts a default value and an ``Optional`` value of the
+        same type and returns the default value if the optional value is
+        ``None`` or the optional value otherwise.
+        """
+        assert_that(fromOptional(1, None), equal_to(1))
+        assert_that(fromOptional(2, 2), equal_to(2))
+
+    def test_discardResult(self) -> None:
+        """
+        ``discardResult`` accepts an awaitable and returns a ``Deferred`` that
+        fires with ``None`` after the awaitable completes.
+        """
+        a: Deferred[str] = Deferred()
+        d = discardResult(a)
+        self.assertNoResult(d)
+        a.callback("result")
+        assert_that(self.successResultOf(d), none())
+
+    def test_sequence(self):
+        """
+        ``sequence`` accepts two awaitables and returns an awaitable that waits
+        for the first one to complete and then completes with the result of
+        the second one.
+        """
+        a: Deferred[str] = Deferred()
+        b: Deferred[int] = Deferred()
+        c = Deferred.fromCoroutine(sequence(a, b))
+        b.callback(42)
+        self.assertNoResult(c)
+        a.callback("hello")
+        assert_that(self.successResultOf(c), equal_to(42))
+
+    def test_iterateWhile(self):
+        """
+        ``iterateWhile`` executes the actions from its factory until the predicate
+        does not match an action result.
+        """
+        actions: List[Deferred[int]] = [Deferred(), Deferred(), Deferred()]
+
+        def predicate(value):
+            return value != 42
+
+        d: Deferred[int] = Deferred.fromCoroutine(
+            iterateWhile(predicate, list(actions).pop)
+        )
+        # Let the action it is waiting on complete
+        actions.pop().callback(7)
+
+        # It does not match the predicate so it is not done yet.
+        self.assertNoResult(d)
+
+        # Let the action it is waiting on now complete - with the result it
+        # wants.
+        actions.pop().callback(42)
+
+        assert_that(self.successResultOf(d), equal_to(42))
+
+    def test_countingCalls(self):
+        """
+        ``countingCalls`` decorates a function so that it is called with an
+        increasing counter and passes the return value through.
+        """
+
+        @countingCalls
+        def target(n: int) -> int:
+            return n + 1
+
+        for expected in range(1, 10):
+            assert_that(target(), equal_to(expected))
+
+
+class WorkerPoolBroken(Exception):
+    """
+    An exception for ``StartedWorkerPoolBroken`` to fail with to allow tests
+    to exercise exception code paths.
+    """
+
+
+class StartedWorkerPoolBroken:
+    """
+    A broken, started worker pool.  Its workers cannot run actions.  They
+    always raise an exception.
+    """
+
+    async def run(self, workerAction: WorkerAction) -> None:
+        raise WorkerPoolBroken()
+
+    async def join(self) -> None:
+        return None
+
+
+@define
+class BrokenWorkerPool:
+    """
+    A worker pool that has workers with a broken ``run`` method.
+    """
+
+    _config: WorkerPoolConfig
+
+    async def start(
+        self, reactor: interfaces.IReactorProcess
+    ) -> StartedWorkerPoolBroken:
+        return StartedWorkerPoolBroken()
+
+
+class _LocalWorker:
+    """
+    A L{Worker} that runs tests in this process in the usual way.
+
+    This is a test double for L{LocalWorkerAMP} which allows testing worker
+    pool logic without sending tests over an AMP connection to be run
+    somewhere else..
+    """
+
+    async def run(self, case: TestCase, result: TestResult) -> None:
+        """
+        Directly run C{case} in the usual way.
+        """
+        TrialSuite([case]).run(result)
+
+
+class WorkerBroken(Exception):
+    """
+    A worker tried to run a test case but the worker is broken.
+    """
+
+
+class _BrokenLocalWorker:
+    """
+    A L{Worker} that always fails to run test cases.
+    """
+
+    async def run(self, case: TestCase, result: TestResult) -> None:
+        """
+        Raise an exception instead of running C{case}.
+        """
+        raise WorkerBroken()
+
+
+@define
+class StartedLocalWorkerPool:
+    """
+    A started L{LocalWorkerPool}.
+    """
+
+    workingDirectory: FilePath
+    workers: List[Worker]
+    _stopped: Deferred
+
+    async def run(self, workerAction: WorkerAction) -> None:
+        """
+        Run the action with each local worker.
+        """
+        for worker in self.workers:
+            await workerAction(worker)
+
+    async def join(self):
+        await self._stopped
+
+
+@define
+class LocalWorkerPool:
+    """
+    Implement a worker pool that runs tests in-process instead of in child
+    processes.
+    """
+
+    _config: WorkerPoolConfig
+    _started: List[StartedLocalWorkerPool] = field(default=Factory(list))
+    _autostop: bool = False
+    _workerFactory: Callable[[], Worker] = _LocalWorker
+
+    async def start(
+        self, reactor: interfaces.IReactorProcess
+    ) -> StartedLocalWorkerPool:
+        workers = [self._workerFactory() for i in range(self._config.numWorkers)]
+        started = StartedLocalWorkerPool(
+            self._config.workingDirectory,
+            workers,
+            (succeed(None) if self._autostop else Deferred()),
+        )
+        self._started.append(started)
+        return started
