@@ -11,9 +11,8 @@ from io import BytesIO, StringIO
 from zope.interface.verify import verifyObject
 
 from twisted.internet.defer import fail, succeed
-from twisted.internet.error import ConnectionDone
+from twisted.internet.error import ProcessDone
 from twisted.internet.interfaces import IAddress, ITransport
-from twisted.internet.main import CONNECTION_DONE
 from twisted.protocols.amp import AMP
 from twisted.python.failure import Failure
 from twisted.python.filepath import FilePath
@@ -25,6 +24,7 @@ from twisted.trial._dist.worker import (
     LocalWorker,
     LocalWorkerAMP,
     LocalWorkerTransport,
+    NotRunning,
     WorkerException,
     WorkerProtocol,
 )
@@ -325,10 +325,36 @@ class LocalWorkerTests(TestCase):
         """
         worker = LocalWorker(*args, **kwargs)
         worker.makeConnection(FakeTransport())
-        self.addCleanup(worker._testLog.close)
         self.addCleanup(worker._outLog.close)
         self.addCleanup(worker._errLog.close)
         return worker
+
+    def test_exitBeforeConnected(self):
+        """
+        L{LocalWorker.exit} fails with L{NotRunning} if it is called before the
+        protocol is connected to a transport.
+        """
+        worker = LocalWorker(
+            SpyDataLocalWorkerAMP(), FilePath(self.mktemp()), StringIO()
+        )
+        self.failureResultOf(worker.exit(), NotRunning)
+
+    def test_exitAfterDisconnected(self):
+        """
+        L{LocalWorker.exit} fails with L{NotRunning} if it is called after the the
+        protocol is disconnected from its transport.
+        """
+        worker = self.tidyLocalWorker(
+            SpyDataLocalWorkerAMP(), FilePath(self.mktemp()), StringIO()
+        )
+        worker.processEnded(Failure(ProcessDone(0)))
+        # Since we're not calling exit until after the process has ended, it
+        # won't consume the ProcessDone failure on the internal `endDeferred`.
+        # Swallow it here.
+        self.failureResultOf(worker.endDeferred, ProcessDone)
+
+        # Now assert that exit behaves.
+        self.failureResultOf(worker.exit(), NotRunning)
 
     def test_childDataReceived(self):
         """
@@ -337,7 +363,7 @@ class LocalWorkerTests(TestCase):
         C{ProcessProtocol.childDataReceived}.
         """
         localWorker = self.tidyLocalWorker(
-            SpyDataLocalWorkerAMP(), self.mktemp(), "test.log"
+            SpyDataLocalWorkerAMP(), FilePath(self.mktemp()), "test.log"
         )
         localWorker._outLog = BytesIO()
         localWorker.childDataReceived(4, b"foo")
@@ -351,25 +377,24 @@ class LocalWorkerTests(TestCase):
         """
         amp = SpyDataLocalWorkerAMP()
         tempDir = FilePath(self.mktemp())
-        logFile = tempDir.child("test.log")
+        tempDir.makedirs()
+        logPath = tempDir.child("test.log")
 
-        worker = LocalWorker(amp, tempDir.path, "test.log")
-        worker.makeConnection(FakeTransport())
-        self.addCleanup(worker._outLog.close)
-        self.addCleanup(worker._errLog.close)
+        with open(logPath.path, "wt") as logFile:
+            worker = LocalWorker(amp, tempDir, logFile)
+            worker.makeConnection(FakeTransport())
+            self.addCleanup(worker._outLog.close)
+            self.addCleanup(worker._errLog.close)
 
-        expected = "Here comes the \N{sun}!"
-        try:
+            expected = "Here comes the \N{sun}!"
             amp.testWrite(expected)
-        finally:
-            worker._testLog.close()
 
         self.assertEqual(
             # os.linesep is the local newline.
             (expected + os.linesep),
             # getContent reads in binary mode so we'll see the bytes that
             # actually ended up in the file.
-            logFile.getContent().decode("utf-8"),
+            logPath.getContent().decode("utf-8"),
         )
 
     def test_outReceived(self):
@@ -378,7 +403,7 @@ class LocalWorkerTests(TestCase):
         file.
         """
         localWorker = self.tidyLocalWorker(
-            SpyDataLocalWorkerAMP(), self.mktemp(), "test.log"
+            SpyDataLocalWorkerAMP(), FilePath(self.mktemp()), "test.log"
         )
         localWorker._outLog = BytesIO()
         data = b"The quick brown fox jumps over the lazy dog"
@@ -391,7 +416,7 @@ class LocalWorkerTests(TestCase):
         file.
         """
         localWorker = self.tidyLocalWorker(
-            SpyDataLocalWorkerAMP(), self.mktemp(), "test.log"
+            SpyDataLocalWorkerAMP(), FilePath(self.mktemp()), "test.log"
         )
         localWorker._errLog = BytesIO()
         data = b"The quick brown fox jumps over the lazy dog"
@@ -433,16 +458,15 @@ class LocalWorkerTests(TestCase):
 
     def test_connectionLost(self):
         """
-        L{LocalWorker.connectionLost} closes the log streams.
+        L{LocalWorker.connectionLost} closes the per-worker log streams.
         """
 
         localWorker = self.tidyLocalWorker(
-            SpyDataLocalWorkerAMP(), self.mktemp(), "test.log"
+            SpyDataLocalWorkerAMP(), FilePath(self.mktemp()), "test.log"
         )
         localWorker.connectionLost(None)
         self.assertTrue(localWorker._outLog.closed)
         self.assertTrue(localWorker._errLog.closed)
-        self.assertTrue(localWorker._testLog.closed)
 
     def test_processEnded(self):
         """
@@ -451,14 +475,13 @@ class LocalWorkerTests(TestCase):
         """
         transport = FakeTransport()
         protocol = SpyDataLocalWorkerAMP()
-        localWorker = LocalWorker(protocol, self.mktemp(), "test.log")
+        localWorker = LocalWorker(protocol, FilePath(self.mktemp()), "test.log")
         localWorker.makeConnection(transport)
-        localWorker.processEnded(Failure(CONNECTION_DONE))
+        localWorker.processEnded(Failure(ProcessDone(0)))
         self.assertTrue(localWorker._outLog.closed)
         self.assertTrue(localWorker._errLog.closed)
-        self.assertTrue(localWorker._testLog.closed)
         self.assertIdentical(None, protocol.transport)
-        return self.assertFailure(localWorker.endDeferred, ConnectionDone)
+        return self.assertFailure(localWorker.endDeferred, ProcessDone)
 
     def test_addresses(self):
         """
@@ -488,6 +511,5 @@ class LocalWorkerTests(TestCase):
 
         protocol = SpyDataLocalWorkerAMP()
         protocol.callRemote = failCallRemote
-        self.tidyLocalWorker(protocol, self.mktemp(), "test.log")
-
+        self.tidyLocalWorker(protocol, FilePath(self.mktemp()), "test.log")
         self.assertEqual([], self.flushLoggedErrors(RuntimeError))
